requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==5.0.4
importlib-metadata==5.0.0
Markdown==3.4.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
zipp==3.10.0
